fix(fuse): preserve lineage proofs through spill uploads and cancellation
